+++ .gitignore
@@ -1,3 +1,5 @@
 node_modules/
 .vscode/
 .env
+logfile.log
+uncaughtExceptions.log
