+++ api/controllers/article.js
@@ -1,0 +1,7 @@
+
+module.exports = {
+    getAllArticles:  async (req, res) => {
+        const articles = await Article.find();
+        res.send(articles);
+    }
+}
