+++ api/routes/auth.js
@@ -1,0 +1,7 @@
+const express = require('express');
+const router = express.Router();
+const authController = require('../controllers/auth');
+
+router.post('/', authController.singUp);
+
+module.exports = router;
