+++ logfile.log
@@ -30,3 +30,8 @@
 {"level":"error","message":"Parameter \"filter\" to find() must be an object, got [object Promise] Parameter \"filter\" to find() must be an object, got [object Promise]","stack":"ObjectParameterError: Parameter \"filter\" to find() must be an object, got [object Promise]\n    at model.Query.Query.find (C:\\Users\\Никита\\Desktop\\test_project\\node_modules\\mongoose\\lib\\query.js:1964:16)\n    at Function.find (C:\\Users\\Никита\\Desktop\\test_project\\node_modules\\mongoose\\lib\\model.js:2052:13)\n    at C:\\Users\\Никита\\Desktop\\test_project\\routes\\channels.js:40:38\n    at newFn (C:\\Users\\Никита\\Desktop\\test_project\\node_modules\\express-async-errors\\index.js:16:20)\n    at Layer.handle [as handle_request] (C:\\Users\\Никита\\Desktop\\test_project\\node_modules\\express\\lib\\router\\layer.js:95:5)\n    at next (C:\\Users\\Никита\\Desktop\\test_project\\node_modules\\express\\lib\\router\\route.js:137:13)\n    at Route.dispatch (C:\\Users\\Никита\\Desktop\\test_project\\node_modules\\express\\lib\\router\\route.js:112:3)\n    at newFn (C:\\Users\\Никита\\Desktop\\test_project\\node_modules\\express-async-errors\\index.js:16:20)\n    at Layer.handle [as handle_request] (C:\\Users\\Никита\\Desktop\\test_project\\node_modules\\express\\lib\\router\\layer.js:95:5)\n    at C:\\Users\\Никита\\Desktop\\test_project\\node_modules\\express\\lib\\router\\index.js:281:22\n    at Function.process_params (C:\\Users\\Никита\\Desktop\\test_project\\node_modules\\express\\lib\\router\\index.js:335:12)\n    at next (C:\\Users\\Никита\\Desktop\\test_project\\node_modules\\express\\lib\\router\\index.js:275:10)\n    at Function.handle (C:\\Users\\Никита\\Desktop\\test_project\\node_modules\\express\\lib\\router\\index.js:174:3)\n    at router (C:\\Users\\Никита\\Desktop\\test_project\\node_modules\\express\\lib\\router\\index.js:47:12)\n    at newFn (C:\\Users\\Никита\\Desktop\\test_project\\node_modules\\express-async-errors\\index.js:16:20)\n    at Layer.handle [as handle_request] (C:\\Users\\Никита\\Desktop\\test_project\\node_modules\\express\\lib\\router\\layer.js:95:5)"}
 {"message":"Listening on port 3000","level":"info"}
 {"message":"Connected to MongoDB...","level":"info"}
+{"error":{"code":"MODULE_NOT_FOUND","requireStack":["C:\\Users\\Никита\\Desktop\\test_project\\startup\\routes.js","C:\\Users\\Никита\\Desktop\\test_project\\index.js"]},"level":"error","message":"uncaughtException: Cannot find module '../routes/channels'\nRequire stack:\n- C:\\Users\\Никита\\Desktop\\test_project\\startup\\routes.js\n- C:\\Users\\Никита\\Desktop\\test_project\\index.js\nError: Cannot find module '../routes/channels'\nRequire stack:\n- C:\\Users\\Никита\\Desktop\\test_project\\startup\\routes.js\n- C:\\Users\\Никита\\Desktop\\test_project\\index.js\n    at Function.Module._resolveFilename (internal/modules/cjs/loader.js:880:15)\n    at Function.Module._load (internal/modules/cjs/loader.js:725:27)\n    at Module.require (internal/modules/cjs/loader.js:952:19)\n    at require (internal/modules/cjs/helpers.js:88:18)\n    at Object.<anonymous> (C:\\Users\\Никита\\Desktop\\test_project\\startup\\routes.js:2:18)\n    at Module._compile (internal/modules/cjs/loader.js:1063:30)\n    at Object.Module._extensions..js (internal/modules/cjs/loader.js:1092:10)\n    at Module.load (internal/modules/cjs/loader.js:928:32)\n    at Function.Module._load (internal/modules/cjs/loader.js:769:14)\n    at Module.require (internal/modules/cjs/loader.js:952:19)","stack":"Error: Cannot find module '../routes/channels'\nRequire stack:\n- C:\\Users\\Никита\\Desktop\\test_project\\startup\\routes.js\n- C:\\Users\\Никита\\Desktop\\test_project\\index.js\n    at Function.Module._resolveFilename (internal/modules/cjs/loader.js:880:15)\n    at Function.Module._load (internal/modules/cjs/loader.js:725:27)\n    at Module.require (internal/modules/cjs/loader.js:952:19)\n    at require (internal/modules/cjs/helpers.js:88:18)\n    at Object.<anonymous> (C:\\Users\\Никита\\Desktop\\test_project\\startup\\routes.js:2:18)\n    at Module._compile (internal/modules/cjs/loader.js:1063:30)\n    at Object.Module._extensions..js (internal/modules/cjs/loader.js:1092:10)\n    at Module.load (internal/modules/cjs/loader.js:928:32)\n    at Function.Module._load (internal/modules/cjs/loader.js:769:14)\n    at Module.require (internal/modules/cjs/loader.js:952:19)","exception":true,"date":"Tue Jan 26 2021 17:55:01 GMT+0200 (Восточная Европа, стандартное время)","process":{"pid":11532,"uid":null,"gid":null,"cwd":"C:\\Users\\Никита\\Desktop\\test_project","execPath":"C:\\Program Files\\nodejs\\node.exe","version":"v14.15.1","argv":["C:\\Program Files\\nodejs\\node.exe","C:\\Users\\Никита\\Desktop\\test_project\\index.js"],"memoryUsage":{"rss":33734656,"heapTotal":16060416,"heapUsed":9090256,"external":1416103,"arrayBuffers":109336}},"os":{"loadavg":[0,0,0],"uptime":1123335},"trace":[{"column":15,"file":"internal/modules/cjs/loader.js","function":"Module._resolveFilename","line":880,"method":"_resolveFilename","native":false},{"column":27,"file":"internal/modules/cjs/loader.js","function":"Module._load","line":725,"method":"_load","native":false},{"column":19,"file":"internal/modules/cjs/loader.js","function":"Module.require","line":952,"method":"require","native":false},{"column":18,"file":"internal/modules/cjs/helpers.js","function":"require","line":88,"method":null,"native":false},{"column":18,"file":"C:\\Users\\Никита\\Desktop\\test_project\\startup\\routes.js","function":null,"line":2,"method":null,"native":false},{"column":30,"file":"internal/modules/cjs/loader.js","function":"Module._compile","line":1063,"method":"_compile","native":false},{"column":10,"file":"internal/modules/cjs/loader.js","function":"Module._extensions..js","line":1092,"method":".js","native":false},{"column":32,"file":"internal/modules/cjs/loader.js","function":"Module.load","line":928,"method":"load","native":false},{"column":14,"file":"internal/modules/cjs/loader.js","function":"Module._load","line":769,"method":"_load","native":false},{"column":19,"file":"internal/modules/cjs/loader.js","function":"Module.require","line":952,"method":"require","native":false}]}
+{"message":"Listening on port 3000","level":"info"}
+{"message":"Connected to MongoDB...","level":"info"}
+{"message":"Listening on port 3000","level":"info"}
+{"message":"Connected to MongoDB...","level":"info"}
